+++ packages/orq-rc/src/models/components/evaluatorresponsefunction.ts
@@ -856,8 +856,8 @@ export const EvaluatorResponseFunction$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-22T22:32:40.794Z"),
-  updated: z.string().default("2026-07-22T22:32:40.794Z"),
+  created: z.string().default("2026-07-23T02:27:42.927Z"),
+  updated: z.string().default("2026-07-23T02:27:42.927Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("function_eval"),