+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-22T22:32:36.266Z",
+    "2026-07-23T02:27:35.252Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {