+++ packages/orq-rc/RELEASES.md
@@ -11109,3 +11109,13 @@ Based on:
 - [typescript v4.13.0-rc.0] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.0 - packages/orq-rc
+
+## 2026-07-23 02:30:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.1 - packages/orq-rc